Framework for module with testing helper functions
